+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Changed
 * Update the eval workflow to intialize agent only once and use different thread ids for each question / conversation
+* Each eval run has it's own directory in `evals/output/` containing `agent_response.csv`, `failed_response.csv` and `metadata.csv`
+* Refactor evaluation workflow to update eval records in place
+* Refactor evaluation workflow to also generate a report `.txt` file at the end of the run
 
 ### Added
 * Add RAGAS metrics to evaluation workflow - `rouge_f1_score`, `factual_correctness_f1_score`, `answer_relevancy_score`